+++ packages/ui/src/LoadingIcon/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 const LoadingIcon = (
-  style: Pick<React.CSSProperties, "fontSize" | "color">
+  style: Pick<React.CSSProperties, "fontSize" | "color">,
 ) => {
   return <div data-testid="loading" className="loading" style={style} />;
 };