+++ packages/user/src/components/Invitation/InvitationModal.tsx
@@ -4,20 +4,28 @@ import { Dialog } from "primereact/dialog";
 import { IconType } from "primereact/utils";
 import React, { useState } from "react";
 
+import {
+  AddInvitationResponse,
+  InvitationAppOption,
+  InvitationRoleOption,
+} from "@/types";
+
 import { InvitationForm } from "./InvitationForm";
 
-import type { InvitationPayload } from "../../types";
-
 interface Properties {
-  handleSubmit: (data: InvitationPayload) => void;
-  loading?: boolean;
+  apps?: InvitationAppOption[];
+  roles?: InvitationRoleOption[];
   buttonIcon?: IconType<ButtonProps>;
+  prepareData?: (data: any) => any;
+  onSubmitted?: (response: AddInvitationResponse) => void;
 }
 
 export const InvitationModal = ({
-  handleSubmit,
-  loading,
+  apps,
   buttonIcon,
+  roles,
+  prepareData,
+  onSubmitted,
 }: Properties) => {
   const { t } = useTranslation("user");
   const [modalVisible, setModalVisible] = useState<boolean>(false);
@@ -38,11 +46,19 @@ export const InvitationModal = ({
         resizable={false}
       >
         <InvitationForm
-          handleSubmit={handleSubmit}
+          apps={apps}
+          roles={roles}
           onCancel={() => {
             setModalVisible(false);
           }}
-          loading={loading}
+          onSubmitted={(data) => {
+            if (onSubmitted) {
+              onSubmitted(data);
+            }
+
+            setModalVisible(false);
+          }}
+          prepareData={prepareData}
         />
       </Dialog>
     </div>