+++ packages/user/src/components/Invitation/InvitationsTable.tsx
@@ -10,7 +10,11 @@ import { InvitationActions } from "./InvitationActions";
 
 import { InvitationModal } from ".";
 
-import type { InvitationPayload } from "../../types";
+import type {
+  AddInvitationResponse,
+  InvitationAppOption,
+  InvitationRoleOption,
+} from "@/types";
 
 export type InvitationsTableProperties = {
   id?: string;
@@ -20,10 +24,13 @@ export type InvitationsTableProperties = {
   showInviteAction?: boolean;
   totalRecords?: number;
   invitations: Array<object>;
+  apps?: Array<InvitationAppOption>;
+  roles?: Array<InvitationRoleOption>;
   fetchInvitations: (arguments_?: any) => void;
-  handleInvitationSubmit?: (data: InvitationPayload) => void;
-  handleInvitationResend?: (data: any) => void;
-  handleInvitationRevoke?: (data: any) => void;
+  onInvitationAdded?: (response: AddInvitationResponse) => void;
+  onInvitationResent?: (data: any) => void;
+  onInvitationRevoked?: (data: any) => void;
+  prepareInvitationData?: (data: any) => any;
   inviteButtonIcon?: IconType<ButtonProps>;
 };
 
@@ -35,10 +42,13 @@ export const InvitationsTable = ({
   showInviteAction = true,
   totalRecords = 0,
   invitations,
+  apps,
+  roles,
   fetchInvitations,
-  handleInvitationSubmit,
-  handleInvitationResend,
-  handleInvitationRevoke,
+  onInvitationAdded,
+  onInvitationResent,
+  onInvitationRevoked,
+  prepareInvitationData,
   inviteButtonIcon,
 }: InvitationsTableProperties) => {
   const { t } = useTranslation("user");
@@ -77,7 +87,11 @@ export const InvitationsTable = ({
       field: "invitedBy",
       header: t("invitations.table.defaultColumns.invitedBy"),
       body: (data) => {
-        return data.invitedBy;
+        return data.invitedBy ? (
+          `${data.invitedBy.givenName} ${data.invitedBy.surname}`
+        ) : (
+          <code>&#8212;</code>
+        );
       },
       align: "center",
     },
@@ -98,9 +112,9 @@ export const InvitationsTable = ({
         return (
           <>
             <InvitationActions
-              handleInvitationResend={handleInvitationResend}
-              handleInvitationRevoke={handleInvitationRevoke}
-              data={data}
+              onInvitationResent={onInvitationResent}
+              onInvitationRevoked={onInvitationRevoked}
+              invitation={data}
             />
           </>
         );
@@ -114,13 +128,15 @@ export const InvitationsTable = ({
   };
 
   const renderHeader = () => {
-    if (showInviteAction && handleInvitationSubmit) {
+    if (showInviteAction) {
       return (
         <div className="table-actions">
           <InvitationModal
-            handleSubmit={handleInvitationSubmit}
-            loading={loading}
+            apps={apps}
+            roles={roles}
             buttonIcon={inviteButtonIcon}
+            onSubmitted={onInvitationAdded}
+            prepareData={prepareInvitationData}
           />
         </div>
       );