+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# [0.14.0](https://github.com/dzangolab/react/compare/v0.13.2...v0.14.0) (2023-08-07)
+
+
+### Breaking changes
+* **user:** handle invitation operations including add, resend, revoke and accept internally ([#488](https://github.com/dzangolab/react/issues/488))
+
+
+### Features
+
+* **user:** add app support for invitations ([#489](https://github.com/dzangolab/react/issues/489)) ([63786d6](https://github.com/dzangolab/react/commit/63786d6863e317ab1b5a6f006e63f415b1765089))
+* **user:** add components and related api for all invitation features ([#488](https://github.com/dzangolab/react/issues/488)) ([adc040f](https://github.com/dzangolab/react/commit/adc040f244153bbc2f27f8c7db0109cfca4d9c61))
+
+
+
+## [0.13.2](https://github.com/dzangolab/react/compare/v0.13.1...v0.13.2) (2023-08-03)
+
+
+### Bug Fixes
+
+* **deps:** update dependency eslint-plugin-react to v7.33.0 ([#471](https://github.com/dzangolab/react/issues/471)) ([2852c93](https://github.com/dzangolab/react/commit/2852c93c8bb18d4132a186fcc47bcc0943db48bc))
+* remove unnecessary space above the header in user table ([#478](https://github.com/dzangolab/react/issues/478)) ([ed63700](https://github.com/dzangolab/react/commit/ed637003a65f54183c778b6584164a7fe6ee2bf4))
+* update invited by column to match the backend response ([#481](https://github.com/dzangolab/react/issues/481)) ([c929601](https://github.com/dzangolab/react/commit/c929601b2c4f1ec08472cac591c295c070d2b12c))
+
+
+
 ## [0.13.1](https://github.com/dzangolab/react/compare/v0.13.0...v0.13.1) (2023-07-26)